+++ Sloop/Abstractions/IDbCacheMigrator.cs
@@ -1,0 +1,6 @@
+namespace Sloop.Abstractions;
+
+public interface IDbCacheMigrator
+{
+    Task MigrateAsync(CancellationToken ct = default);
+}