+++ Sloop/Abstractions/IDbConnectionFactory.cs
@@ -1,4 +1,4 @@
-namespace Sloop.Interfaces;
+namespace Sloop.Abstractions;
 
 using Npgsql;
 
@@ -10,5 +10,5 @@ public interface IDbConnectionFactory
     /// <summary>
     ///     Creates and returns a new open <see cref="NpgsqlConnection" />.
     /// </summary>
-    Task<NpgsqlConnection> Create(CancellationToken token = default);
+    Task<NpgsqlConnection> Create(CancellationToken ct = default);
 }