+++ Sloop/Commands/GetItemCommand.cs
@@ -1,6 +1,7 @@
 namespace Sloop.Commands;
 
-using Interfaces;
+using Abstractions;
+using Core;
 using Microsoft.Extensions.Options;
 using Npgsql;
 
@@ -34,7 +35,7 @@ public class GetItemCommand : IDbCacheCommand<GetItemArgs, byte[]?>
 
         cmd.CommandText =
             $"""
-             UPDATE "{_options.SchemaName}"."{_options.TableName}"
+             UPDATE {_options.GetQualifiedTableName()}
              SET expires_at = LEAST(now() + sliding_interval, absolute_expiry)
              WHERE key = @key
                AND (expires_at IS NULL OR expires_at > now())